requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-multipart==0.0.22
jinja2==3.1.6
opencv-python-headless==4.13.0.92
numpy==2.4.3
python-dotenv==1.2.1
